put organism class and eval functions into own files, add default+log fitness plotting
